# source/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/b2b-product": {
            "get": {
                "description": "Get all b2b products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "b2b-product"
                ],
                "summary": "Get all b2b products",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "role",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    }
                }
            },
            "post": {
                "description": "Create b2b product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "b2b-product"
                ],
                "summary": "Create b2b product",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.RequestB2BProduct"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    }
                }
            }
        },
        "/b2b-product/bulk-update": {
            "put": {
                "description": "Bulk update b2b products",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "b2b-product"
                ],
                "summary": "Bulk update b2b products",
                "parameters": [
                    {
                        "type": "file",
                        "description": "CSV File Structure:\n Header: id,name,price,vital-signs,visus,physical-test,injection,phlebotomy,specimen-mandiri,pap-smear,rontgen,ekg,audiometri,spirometri,treadmill,usg-abdomen,usg-mammae;\n Example Value: 1,MCU Basic,10000,1,1,1,1,1,1,1,1,1,1,1,1,1,1 \n Note: if you want to create new product, you can leave the id as empty, else if you want to update the product, you can fill the id with the product id",
                        "name": "csv_file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    }
                }
            }
        },
        "/b2b-product/generate-slugs": {
            "post": {
                "description": "Generate slugs for products where slug is null",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "b2b-product"
                ],
                "summary": "Generate slugs for products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    }
                }
            }
        },
...
        "/b2b-product/{b2b_product_id}": {
            "get": {
                "description": "Get b2b product by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "b2b-product"
                ],
                "summary": "Get b2b product by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "B2B Product ID",
                        "name": "b2b_product_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    }
                }
            },
            "put": {
                "description": "Update b2b product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "b2b-product"
                ],
                "summary": "Update b2b product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "B2B Product ID",
                        "name": "b2b_product_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.RequestB2BProduct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete b2b product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "b2b-product"
                ],
                "summary": "Delete b2b product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "B2B Product ID",
                        "name": "b2b_product_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "web.ErrorResponse": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "web.RequestB2BProduct": {
            "type": "object",
            "required": [
                "name",
                "price"
            ],
            "properties": {
                "examination_checklist": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "injection",
                        "phlebotomy"
                    ]
                },
                "id": {
                    "type": "integer",
                    "default": 0,
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Product Name"
                },
                "price": {
                    "type": "integer",
                    "example": 10000000
                }
            }
        },
...
        "web.ResponseStandard": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.ErrorResponse"
                    }
                },
                "fulfilled": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "description": "Meta       any `json:\"meta\"`\nJsonApi    any `json:\"jsonapi\"`\nLinks      any `json:\"links\"`\nIncluded   any `json:\"included\"`\nSelf       any `json:\"self\"`\nRelated    any `json:\"related\"`"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        }
    }
}